Now use "Location and GeoPoint"
Fix up a few missing items in location
