+++ Caboodle/Geocoding/Geocoding.netstandard.cs
@@ -5,11 +5,11 @@ namespace Microsoft.Caboodle
 {
 	public partial class Geocoding
     {
-        public static Task<IEnumerable<Address>> GetAddressesAsync(double latitude, double longitude) =>
+        public static Task<IEnumerable<Location>> GetLocationsAsync(double latitude, double longitude) =>
             throw new NotImplentedInReferenceAssembly();
 
 
-        public static Task<IEnumerable<Position>> GetPositionsAsync(string address) =>
+        public static Task<IEnumerable<GeoPoint>> GetGeoPointsAsync(string address) =>
             throw new NotImplentedInReferenceAssembly();
     }
 }
