--- Caboodle/Types/PositionExtensions.ios.cs
@@ -1,7 +1,0 @@
-﻿namespace Microsoft.Caboodle
-{
-	public static partial class PositionExtensions
-    {
-		
-	}
-}
